- cleanup dead code #741 - team manager revamped with slotted reservations list and members lists #723 - rooming manager (WIP)
